+++ common/src/main/scala/com/gu/recipeasy/models/CuratedRecipe.scala
@@ -100,10 +100,9 @@ object Tag {
 }
 
 object CuratedRecipe {
-
   import CuratedRecipeDB._
 
-  def toCuratedRecipeDB(cr: CuratedRecipe): CuratedRecipeDB = {
+  def toDBModel(cr: CuratedRecipe): CuratedRecipeDB = {
     transform[CuratedRecipe, CuratedRecipeDB](
       cr,
       "tags" -> getTagNames(cr.tags)
